+++ pandapower/convert_format.py
@@ -232,14 +232,14 @@ def _add_missing_columns(net):
                 net.controller.at[ctrl.name, 'initial_run'] = ctrl['object'].initial_powerflow
 
     # distributed slack
-    if "contribution_factor" not in net.ext_grid:
-        net.ext_grid['contribution_factor'] = 1.0
+    if "slack_weight" not in net.ext_grid:
+        net.ext_grid['slack_weight'] = 1.0
 
-    if "contribution_factor" not in net.gen:
-        net.gen['contribution_factor'] = 0.0
+    if "slack_weight" not in net.gen:
+        net.gen['slack_weight'] = 0.0
 
-    if "contribution_factor" not in net.xward:
-        net.xward['contribution_factor'] = 0.0
+    if "slack_weight" not in net.xward:
+        net.xward['slack_weight'] = 0.0
 
 
 def _update_trafo_type_parameter_names(net):